Move create thread id to utils
[ghstack-poisoned]
