+++ apeiron/agent.py
@@ -13,35 +13,9 @@ from .chat_message_histories.discord import DiscordChannelChatMessageHistory
 from .chat_models import create_chat_model
 from .toolkits.discord.toolkit import DiscordToolkit
 from .tools.discord.utils import is_client_user
-from .utils import parse_feature_gates
+from .utils import parse_feature_gates, create_thread_id
 
 logger = logging.getLogger(__name__)
-
-
-def create_thread_id(message: discord.Message):
-    """Create a thread ID from a Discord message."""
-    if message.guild is None:
-        return "/".join([
-            "guild",
-            "__private__",
-            "channel",
-            str(message.author.id),
-        ])
-    if message.thread is None:
-        return "/".join([
-            "guild",
-            str(message.guild.id),
-            "channel",
-            str(message.channel.id),
-        ])
-    return "/".join([
-        "guild",
-        str(message.guild.id),
-        "channel",
-        str(message.channel.id),
-        "thread",
-        str(message.thread.id),
-    ])
 
 
 def create_bot(bot: discord.Bot, model: BaseChatModel, pregel: Pregel):
